feat: add custom hook - useVariables
